test file for "Measurements of Cache Effects"
File for testing the difference of cycles/list of elements between the random array and the sequential array. So far i created the linked list, the random array, the sequential. i randomised the node in the random array by randomly shuffle the array by swapping elements in random positions. I'm still working on telling the difference between two arrays.